Nest WorldGuardIntegration in WorldGuardHook
